+++ algorithmic_efficiency/workloads/fastmri/fastmri_jax/ssim.py
